+++ ChangeLog.md
@@ -1,6 +1,19 @@
 Change Log
 ==========
 
+# 5.0.3 (2017-10-03)
+
+### New features
+
+- Index: Add `deleteBy` method and deprecate `deleteByQuery` method
+- PlacesClient: Add `getObject(withID:completionHandler)` method
+
+# 5.0.2 (2017-10-01)
+
+### New features
+
+- Support for Swift 4
+
 ## 4.8.1 (2017-04-18)
 
 ### Miscellaneous changes